refactor(training): application/use_cases.py → 9 subpacotes (Fase 3)
- sessions/{dto,queries,commands}.py — 6 UseCases
- blocks/{dto,queries,commands}.py — 6 UseCases
- wellness/{dto,queries,commands}.py — 6 UseCases
- attendance/{dto,queries,commands}.py — 2 UseCases
- execution/{dto,queries,commands}.py — 5 UseCases
- planning/{dto,queries,commands}.py — 8 UseCases
- communication/{dto,queries,commands}.py — 12 UseCases
- eligibility/{dto,queries,commands}.py — 2 UseCases
- analytics/{dto,queries}.py — 1 UseCase

application/use_cases.py substituído por shim < 165 linhas que
re-exporta todos os 48 *UseCase sem quebrar importações existentes.

domain/policies/feedback_context.py: consolida _feedback_context_type
e _feedback_context_ref_id (removidos de use_cases.py e mappers.py).

test_application_layout.py: garante surface pública do shim, tamanho
de dto.py e ausência de django.conf nos subpacotes.

Testes: 290 passed, 19 skipped (era 270 → +20 novos testes de layout)
